backend/pyproject.toml
[project]
name = "career-copilot-backend"
version = "0.1.0"
description = "Dual-mode CV + JD analyzer built on Google ADK v2."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "uvicorn>=0.40.0",
    "google-adk>=2.0.0a0",
    "google-genai>=0.3.0",
    "litellm>=1.81.0",
    "mcp>=1.0.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
